split messaging subsystem XML parsing and writing
* add MessagingXMLWriter to write the messaging current version in XML
* let the MessagingXXSubsystemParsers handle only the XML parsing of the
  specific versions
