+++ messaging/src/main/java/org/jboss/as/messaging/MessagingXMLWriter.java
@@ -1,0 +1,801 @@
+package org.jboss.as.messaging;
+
+import static org.jboss.as.messaging.CommonAttributes.ACCEPTOR;
+import static org.jboss.as.messaging.CommonAttributes.ADDRESS_SETTING;
+import static org.jboss.as.messaging.CommonAttributes.BINDINGS_DIRECTORY;
+import static org.jboss.as.messaging.CommonAttributes.BROADCAST_GROUP;
+import static org.jboss.as.messaging.CommonAttributes.CONNECTION_FACTORY;
+import static org.jboss.as.messaging.CommonAttributes.CONNECTOR;
+import static org.jboss.as.messaging.CommonAttributes.DEFAULT;
+import static org.jboss.as.messaging.CommonAttributes.DISCOVERY_GROUP;
+import static org.jboss.as.messaging.CommonAttributes.DIVERT;
+import static org.jboss.as.messaging.CommonAttributes.DURABLE;
+import static org.jboss.as.messaging.CommonAttributes.GROUPING_HANDLER;
+import static org.jboss.as.messaging.CommonAttributes.HORNETQ_SERVER;
+import static org.jboss.as.messaging.CommonAttributes.INBOUND_CONFIG;
+import static org.jboss.as.messaging.CommonAttributes.IN_VM_ACCEPTOR;
+import static org.jboss.as.messaging.CommonAttributes.IN_VM_CONNECTOR;
+import static org.jboss.as.messaging.CommonAttributes.JMS_BRIDGE;
+import static org.jboss.as.messaging.CommonAttributes.JMS_CONNECTION_FACTORIES;
+import static org.jboss.as.messaging.CommonAttributes.JMS_DESTINATIONS;
+import static org.jboss.as.messaging.CommonAttributes.JMS_QUEUE;
+import static org.jboss.as.messaging.CommonAttributes.JMS_TOPIC;
+import static org.jboss.as.messaging.CommonAttributes.PARAM;
+import static org.jboss.as.messaging.CommonAttributes.PATH;
+import static org.jboss.as.messaging.CommonAttributes.POOLED_CONNECTION_FACTORY;
+import static org.jboss.as.messaging.CommonAttributes.RELATIVE_TO;
+import static org.jboss.as.messaging.CommonAttributes.REMOTE_ACCEPTOR;
+import static org.jboss.as.messaging.CommonAttributes.REMOTE_CONNECTOR;
+import static org.jboss.as.messaging.CommonAttributes.ROLE;
+import static org.jboss.as.messaging.CommonAttributes.SELECTOR;
+import static org.jboss.as.messaging.CommonAttributes.TRANSACTION;
+import static org.jboss.as.messaging.CommonAttributes.VALUE;
+import static org.jboss.as.messaging.Element.SOURCE;
+import static org.jboss.as.messaging.Element.TARGET;
+import static org.jboss.as.messaging.MessagingMessages.MESSAGES;
+import static org.jboss.as.messaging.Namespace.CURRENT;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.xml.stream.XMLStreamException;
+
+import org.jboss.as.controller.AttributeDefinition;
+import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
+import org.jboss.as.controller.persistence.SubsystemMarshallingContext;
+import org.jboss.as.messaging.jms.ConnectionFactoryAttributes;
+import org.jboss.as.messaging.jms.ConnectionFactoryAttributes.Common;
+import org.jboss.as.messaging.jms.ConnectionFactoryAttributes.Pooled;
+import org.jboss.as.messaging.jms.ConnectorAttribute;
+import org.jboss.as.messaging.jms.JndiEntriesAttribute;
+import org.jboss.as.messaging.jms.bridge.JMSBridgeDefinition;
+import org.jboss.dmr.ModelNode;
+import org.jboss.dmr.ModelType;
+import org.jboss.dmr.Property;
+import org.jboss.staxmapper.XMLElementWriter;
+import org.jboss.staxmapper.XMLExtendedStreamWriter;
+
+public class MessagingXMLWriter implements XMLElementWriter<SubsystemMarshallingContext> {
+
+    static final MessagingXMLWriter INSTANCE = new MessagingXMLWriter();
+
+    private static final char[] NEW_LINE = new char[]{'\n'};
+
+    private MessagingXMLWriter() {
+    }
+
+    @Override
+    public void writeContent(XMLExtendedStreamWriter writer, SubsystemMarshallingContext context) throws XMLStreamException {
+        context.startSubsystemElement(CURRENT.getUriString(), false);
+        final ModelNode node = context.getModelNode();
+
+        if (node.hasDefined(HORNETQ_SERVER)) {
+            final ModelNode servers = node.get(HORNETQ_SERVER);
+            boolean first = true;
+            for (Property prop : servers.asPropertyList()) {
+                writeHornetQServer(writer, prop.getName(), prop.getValue());
+                if (!first) {
+                    writeNewLine(writer);
+                } else {
+                    first = false;
+                }
+            }
+        }
+
+        if (node.hasDefined(JMS_BRIDGE)) {
+            final ModelNode jmsBridges = node.get(JMS_BRIDGE);
+            boolean first = true;for (Property prop : jmsBridges.asPropertyList()) {
+                writeJmsBridge(writer, prop.getName(), prop.getValue());
+                if (!first) {
+                    writeNewLine(writer);
+                } else {
+                    first = false;
+                }
+            }
+        }
+        writer.writeEndElement();
+    }
+
+    private static void writeHornetQServer(final XMLExtendedStreamWriter writer, final String serverName, final ModelNode node) throws XMLStreamException {
+
+        writer.writeStartElement(Element.HORNETQ_SERVER.getLocalName());
+
+        if (!DEFAULT.equals(serverName)) {
+            writer.writeAttribute(Attribute.NAME.getLocalName(), serverName);
+        }
+
+        for (AttributeDefinition simpleAttribute : CommonAttributes.SIMPLE_ROOT_RESOURCE_ATTRIBUTES) {
+            simpleAttribute.marshallAsElement(node, writer);
+        }
+
+        final ModelNode paths = node.get(ModelDescriptionConstants.PATH);
+        if (paths.hasDefined(CommonAttributes.PAGING_DIRECTORY)) {
+            writeDirectory(writer, Element.PAGING_DIRECTORY, node.get(ModelDescriptionConstants.PATH));
+        }
+        if (paths.hasDefined(BINDINGS_DIRECTORY)) {
+            writeDirectory(writer, Element.BINDINGS_DIRECTORY, node.get(ModelDescriptionConstants.PATH));
+        }
+        if (paths.hasDefined(CommonAttributes.JOURNAL_DIRECTORY)) {
+            writeDirectory(writer, Element.JOURNAL_DIRECTORY, node.get(ModelDescriptionConstants.PATH));
+        }
+        if (paths.hasDefined(CommonAttributes.LARGE_MESSAGES_DIRECTORY)) {
+            writeDirectory(writer, Element.LARGE_MESSAGES_DIRECTORY, node.get(ModelDescriptionConstants.PATH));
+        }
+        // New line after the simpler elements
+        writeNewLine(writer);
+
+        writeConnectors(writer, node);
+
+        writeAcceptors(writer, node);
+
+        if (node.hasDefined(BROADCAST_GROUP)) {
+            writeBroadcastGroups(writer, node.get(BROADCAST_GROUP));
+        }
+        if (node.hasDefined(DISCOVERY_GROUP)) {
+            writeDiscoveryGroups(writer, node.get(DISCOVERY_GROUP));
+        }
+        if (node.hasDefined(DIVERT)) {
+            writeDiverts(writer, node.get(DIVERT));
+        }
+        if (node.hasDefined(CommonAttributes.QUEUE)) {
+            writeQueues(writer, node.get(CommonAttributes.QUEUE));
+        }
+        if (node.hasDefined(CommonAttributes.BRIDGE)) {
+            writeBridges(writer, node.get(CommonAttributes.BRIDGE));
+        }
+        if (node.hasDefined(CommonAttributes.CLUSTER_CONNECTION)) {
+            writeClusterConnections(writer, node.get(CommonAttributes.CLUSTER_CONNECTION));
+        }
+
+        if (node.hasDefined(CommonAttributes.GROUPING_HANDLER)) {
+            writeGroupingHandler(writer, node.get(GROUPING_HANDLER));
+        }
+
+        if (node.hasDefined(CommonAttributes.SECURITY_SETTING)) {
+            writeSecuritySettings(writer, node.get(CommonAttributes.SECURITY_SETTING));
+        }
+
+        if (node.hasDefined(ADDRESS_SETTING)) {
+            writeAddressSettings(writer, node.get(ADDRESS_SETTING));
+        }
+
+        if (node.hasDefined(CommonAttributes.CONNECTOR_SERVICE)) {
+            writeConnectorServices(writer, node.get(CommonAttributes.CONNECTOR_SERVICE));
+        }
+
+        if (node.hasDefined(CONNECTION_FACTORY) || node.hasDefined(POOLED_CONNECTION_FACTORY)) {
+            ModelNode cf = node.get(CONNECTION_FACTORY);
+            ModelNode pcf = node.get(POOLED_CONNECTION_FACTORY);
+            boolean hasCf = cf.isDefined() && cf.keys().size() > 0;
+            boolean hasPcf = pcf.isDefined() && pcf.keys().size() > 0;
+            if (hasCf || hasPcf) {
+                writer.writeStartElement(JMS_CONNECTION_FACTORIES);
+                if (hasCf) {
+                    writeConnectionFactories(writer, cf);
+                }
+                if (hasPcf) {
+                    writePooledConnectionFactories(writer, pcf);
+                }
+                writer.writeEndElement();
+                writeNewLine(writer);
+            }
+        }
+
+        if (node.has(JMS_QUEUE) || node.has(JMS_TOPIC)) {
+            ModelNode queue = node.get(JMS_QUEUE);
+            ModelNode topic = node.get(JMS_TOPIC);
+            boolean hasQueue = queue.isDefined() && queue.keys().size() > 0;
+            boolean hasTopic = topic.isDefined() && topic.keys().size() > 0;
+            if (hasQueue || hasTopic) {
+                writer.writeStartElement(JMS_DESTINATIONS);
+                if (hasQueue) {
+                    writeJmsQueues(writer, node.get(JMS_QUEUE));
+                }
+                if (hasTopic) {
+                    writeTopics(writer, node.get(JMS_TOPIC));
+                }
+                writer.writeEndElement();
+            }
+        }
+
+        writer.writeEndElement();
+    }
+
+    private static void writeConnectors(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        if (node.hasDefined(CONNECTOR) || node.hasDefined(REMOTE_CONNECTOR) || node.hasDefined(IN_VM_CONNECTOR)) {
+            writer.writeStartElement(Element.CONNECTORS.getLocalName());
+            if(node.hasDefined(CONNECTOR)) {
+                for(final Property property : node.get(CONNECTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.CONNECTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            if(node.hasDefined(REMOTE_CONNECTOR)) {
+                for(final Property property : node.get(REMOTE_CONNECTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.NETTY_CONNECTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            if(node.hasDefined(IN_VM_CONNECTOR)) {
+                for(final Property property : node.get(IN_VM_CONNECTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.IN_VM_CONNECTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeAcceptors(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        if (node.hasDefined(ACCEPTOR) || node.hasDefined(REMOTE_ACCEPTOR) || node.hasDefined(IN_VM_ACCEPTOR)) {
+            writer.writeStartElement(Element.ACCEPTORS.getLocalName());
+            if(node.hasDefined(ACCEPTOR)) {
+                for(final Property property : node.get(ACCEPTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.ACCEPTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            if(node.hasDefined(REMOTE_ACCEPTOR)) {
+                for(final Property property : node.get(REMOTE_ACCEPTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.NETTY_ACCEPTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            if(node.hasDefined(IN_VM_ACCEPTOR)) {
+                for(final Property property : node.get(IN_VM_ACCEPTOR).asPropertyList()) {
+                    writer.writeStartElement(Element.IN_VM_ACCEPTOR.getLocalName());
+                    writeAcceptorAndConnectorContent(writer, property);
+                    writer.writeEndElement();
+                }
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeAcceptorAndConnectorContent(final XMLExtendedStreamWriter writer, final Property property) throws XMLStreamException {
+        writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+        final ModelNode value = property.getValue();
+
+        if (value.hasDefined(RemoteTransportDefinition.SOCKET_BINDING.getName())) {
+            writeAttribute(writer, Attribute.SOCKET_BINDING, value.get(RemoteTransportDefinition.SOCKET_BINDING.getName()));
+        }
+        if (value.hasDefined(InVMTransportDefinition.SERVER_ID.getName())) {
+            writeAttribute(writer, Attribute.SERVER_ID, value.get(InVMTransportDefinition.SERVER_ID.getName()));
+        }
+
+        CommonAttributes.FACTORY_CLASS.marshallAsElement(value, writer);
+
+        if (value.hasDefined(PARAM)) {
+            for(final Property parameter : value.get(PARAM).asPropertyList()) {
+                writer.writeStartElement(Element.PARAM.getLocalName());
+                writer.writeAttribute(Attribute.KEY.getLocalName(), parameter.getName());
+                writeAttribute(writer, Attribute.VALUE, parameter.getValue().get(VALUE.getName()));
+                writer.writeEndElement();
+            }
+        }
+    }
+
+    private static void writeBroadcastGroups(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.BROADCAST_GROUPS.getLocalName());
+            for(final Property property : properties) {
+                writer.writeStartElement(Element.BROADCAST_GROUP.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                for (AttributeDefinition attribute : BroadcastGroupDefinition.ATTRIBUTES) {
+                    attribute.marshallAsElement(property.getValue(), writer);
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeDiscoveryGroups(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.DISCOVERY_GROUPS.getLocalName());
+            for(final Property property : properties) {
+                writer.writeStartElement(Element.DISCOVERY_GROUP.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                for (AttributeDefinition attribute : DiscoveryGroupDefinition.ATTRIBUTES) {
+                    attribute.marshallAsElement(property.getValue(), writer);
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeDiverts(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.DIVERTS.getLocalName());
+            for(final Property property : properties) {
+                writer.writeStartElement(Element.DIVERT.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                for (AttributeDefinition attribute : DivertDefinition.ATTRIBUTES) {
+                    if (CommonAttributes.FILTER == attribute) {
+                        writeFilter(writer, property.getValue());
+                    } else {
+                        attribute.marshallAsElement(property.getValue(), writer);
+                    }
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeQueues(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.CORE_QUEUES.getLocalName());
+            for (Property queueProp : properties) {
+                writer.writeStartElement(Element.QUEUE.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), queueProp.getName());
+                final ModelNode queue = queueProp.getValue();
+                QueueDefinition.ADDRESS.marshallAsElement(queue, writer);
+                writeFilter(writer, queue);
+                DURABLE.marshallAsElement(queue, writer);
+
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeBridges(XMLExtendedStreamWriter writer, ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.BRIDGES.getLocalName());
+            for(final Property property : node.asPropertyList()) {
+                writer.writeStartElement(Element.BRIDGE.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                final ModelNode bridge = property.getValue();
+                for (AttributeDefinition attribute : BridgeDefinition.ATTRIBUTES) {
+                    if (CommonAttributes.FILTER == attribute) {
+                        writeFilter(writer, property.getValue());
+                    } else if (attribute == CommonAttributes.DISCOVERY_GROUP_NAME) {
+                        if (CommonAttributes.DISCOVERY_GROUP_NAME.isMarshallable(bridge)) {
+                            writer.writeStartElement(Element.DISCOVERY_GROUP_REF.getLocalName());
+                            CommonAttributes.DISCOVERY_GROUP_NAME.marshallAsAttribute(bridge, writer);
+                            writer.writeEndElement();
+                        }
+                    } else {
+                        attribute.marshallAsElement(property.getValue(), writer);
+                    }
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeClusterConnections(XMLExtendedStreamWriter writer, ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.CLUSTER_CONNECTIONS.getLocalName());
+            for(final Property property : node.asPropertyList()) {
+                writer.writeStartElement(Element.CLUSTER_CONNECTION.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                final ModelNode cluster = property.getValue();
+                for (AttributeDefinition attribute : ClusterConnectionDefinition.ATTRIBUTES) {
+                    if (attribute == ClusterConnectionDefinition.ALLOW_DIRECT_CONNECTIONS_ONLY) {
+                        // we nest it in static-connectors
+                        continue;
+                    }
+                    if (attribute == ConnectorRefsAttribute.CLUSTER_CONNECTION_CONNECTORS) {
+                        if (ConnectorRefsAttribute.CLUSTER_CONNECTION_CONNECTORS.isMarshallable(cluster)) {
+                            writer.writeStartElement(Element.STATIC_CONNECTORS.getLocalName());
+                            ClusterConnectionDefinition.ALLOW_DIRECT_CONNECTIONS_ONLY.marshallAsAttribute(cluster, writer);
+                            ConnectorRefsAttribute.CLUSTER_CONNECTION_CONNECTORS.marshallAsElement(cluster, writer);
+                            writer.writeEndElement();
+                        } else if (ClusterConnectionDefinition.ALLOW_DIRECT_CONNECTIONS_ONLY.isMarshallable(cluster)) {
+                            writer.writeEmptyElement(Element.STATIC_CONNECTORS.getLocalName());
+                            ClusterConnectionDefinition.ALLOW_DIRECT_CONNECTIONS_ONLY.marshallAsAttribute(cluster, writer);
+                        }
+                    }
+                    else if (attribute == CommonAttributes.DISCOVERY_GROUP_NAME) {
+                        if (CommonAttributes.DISCOVERY_GROUP_NAME.isMarshallable(cluster)) {
+                            writer.writeStartElement(Element.DISCOVERY_GROUP_REF.getLocalName());
+                            CommonAttributes.DISCOVERY_GROUP_NAME.marshallAsAttribute(cluster, writer);
+                            writer.writeEndElement();
+                        }
+                    } else {
+                        attribute.marshallAsElement(property.getValue(), writer);
+                    }
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeGroupingHandler(XMLExtendedStreamWriter writer, ModelNode node) throws XMLStreamException {
+
+        boolean wroteHandler = false;
+        for (Property handler : node.asPropertyList()) {
+            if (wroteHandler) {
+                throw MESSAGES.multipleChildrenFound(GROUPING_HANDLER);
+            } else {
+                wroteHandler = true;
+            }
+            writer.writeStartElement(Element.GROUPING_HANDLER.getLocalName());
+            writer.writeAttribute(Attribute.NAME.getLocalName(), handler.getName());
+            final ModelNode resourceModel = handler.getValue();
+            for (AttributeDefinition attr : GroupingHandlerDefinition.ATTRIBUTES) {
+                attr.marshallAsElement(resourceModel, writer);
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeFilter(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        if (node.hasDefined(CommonAttributes.FILTER.getName())) {
+            writer.writeEmptyElement(CommonAttributes.FILTER.getXmlName());
+            writer.writeAttribute(CommonAttributes.STRING, node.get(CommonAttributes.FILTER.getName()).asString());
+        }
+    }
+
+    private static void writeDirectory(final XMLExtendedStreamWriter writer, final Element element, final ModelNode node) throws XMLStreamException {
+        final String localName = element.getLocalName();
+        if(node.has(localName)) {
+            final String path = node.get(localName).has(PATH.getName()) ? node.get(localName, PATH.getName()).asString() : null;
+            final String relativeTo = node.get(localName).hasDefined(RELATIVE_TO.getName()) ? node.get(localName, RELATIVE_TO.getName()).asString() : null;
+            if(path != null || relativeTo != null) {
+                writer.writeEmptyElement(localName);
+                if(path != null) writer.writeAttribute(PATH.getName(), path);
+                if(relativeTo != null) writer.writeAttribute(RELATIVE_TO.getName(), relativeTo);
+            }
+        }
+    }
+
+    private static void writeSecuritySettings(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.SECURITY_SETTINGS.getLocalName());
+
+            for (Property matchRoles : properties) {
+                writer.writeStartElement(Element.SECURITY_SETTING.getLocalName());
+                writer.writeAttribute(Attribute.MATCH.getLocalName(), matchRoles.getName());
+
+                if (matchRoles.getValue().hasDefined(ROLE)) {
+
+                    ArrayList<String> send = new ArrayList<String>();
+                    ArrayList<String> consume = new ArrayList<String>();
+                    ArrayList<String> createDurableQueue = new ArrayList<String>();
+                    ArrayList<String> deleteDurableQueue = new ArrayList<String>();
+                    ArrayList<String> createNonDurableQueue = new ArrayList<String>();
+                    ArrayList<String> deleteNonDurableQueue = new ArrayList<String>();
+                    ArrayList<String> manageRoles = new ArrayList<String>();
+
+                    for (Property rolePerms : matchRoles.getValue().get(ROLE).asPropertyList()) {
+                        final String role = rolePerms.getName();
+                        final ModelNode perms = rolePerms.getValue();
+                        if (perms.get(SecurityRoleDefinition.SEND.getName()).asBoolean(false)) {
+                            send.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.CONSUME.getName()).asBoolean(false)) {
+                            consume.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.CREATE_DURABLE_QUEUE.getName()).asBoolean(false)) {
+                            createDurableQueue.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.DELETE_DURABLE_QUEUE.getName()).asBoolean(false)) {
+                            deleteDurableQueue.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.CREATE_NON_DURABLE_QUEUE.getName()).asBoolean(false)) {
+                            createNonDurableQueue.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.DELETE_NON_DURABLE_QUEUE.getName()).asBoolean(false)) {
+                            deleteNonDurableQueue.add(role);
+                        }
+                        if (perms.get(SecurityRoleDefinition.MANAGE.getName()).asBoolean(false)) {
+                            manageRoles.add(role);
+                        }
+                    }
+
+                    writePermission(writer, SecurityRoleDefinition.SEND.getXmlName(), send);
+                    writePermission(writer, SecurityRoleDefinition.CONSUME.getXmlName(), consume);
+                    writePermission(writer, SecurityRoleDefinition.CREATE_DURABLE_QUEUE.getXmlName(), createDurableQueue);
+                    writePermission(writer, SecurityRoleDefinition.DELETE_DURABLE_QUEUE.getXmlName(), deleteDurableQueue);
+                    writePermission(writer, SecurityRoleDefinition.CREATE_NON_DURABLE_QUEUE.getXmlName(), createNonDurableQueue);
+                    writePermission(writer, SecurityRoleDefinition.DELETE_NON_DURABLE_QUEUE.getXmlName(), deleteNonDurableQueue);
+                    writePermission(writer, SecurityRoleDefinition.MANAGE.getXmlName(), manageRoles);
+                }
+
+                writer.writeEndElement();
+            }
+
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writePermission(final XMLExtendedStreamWriter writer, final String type, final List<String> roles) throws XMLStreamException {
+        if (roles.size() == 0) {
+            return;
+        }
+        writer.writeStartElement(Element.PERMISSION_ELEMENT_NAME.getLocalName());
+        StringBuilder sb = new StringBuilder();
+        for (String role : roles) {
+            if (sb.length() > 0) {
+                sb.append(" ");
+            }
+            sb.append(role);
+        }
+        writer.writeAttribute(Attribute.TYPE_ATTR_NAME.getLocalName(), type);
+        writer.writeAttribute(Attribute.ROLES_ATTR_NAME.getLocalName(), sb.toString());
+        writer.writeEndElement();
+    }
+
+    private static void writeAddressSettings(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.ADDRESS_SETTINGS.getLocalName());
+            for (Property matchSetting : properties) {
+                writer.writeStartElement(Element.ADDRESS_SETTING.getLocalName());
+                writer.writeAttribute(Attribute.MATCH.getLocalName(), matchSetting.getName());
+                final ModelNode setting = matchSetting.getValue();
+                CommonAttributes.DEAD_LETTER_ADDRESS.marshallAsElement(setting, writer);
+                CommonAttributes.EXPIRY_ADDRESS.marshallAsElement(setting, writer);
+                AddressSettingDefinition.REDELIVERY_DELAY.marshallAsElement(setting, writer);
+                AddressSettingDefinition.MAX_DELIVERY_ATTEMPTS.marshallAsElement(setting, writer);
+                AddressSettingDefinition.MAX_SIZE_BYTES.marshallAsElement(setting, writer);
+                AddressSettingDefinition.PAGE_SIZE_BYTES.marshallAsElement(setting, writer);
+                AddressSettingDefinition.PAGE_MAX_CACHE_SIZE.marshallAsElement(setting, writer);
+                AddressSettingDefinition.ADDRESS_FULL_MESSAGE_POLICY.marshallAsElement(setting, writer);
+                AddressSettingDefinition.MESSAGE_COUNTER_HISTORY_DAY_LIMIT.marshallAsElement(setting, writer);
+                AddressSettingDefinition.LAST_VALUE_QUEUE.marshallAsElement(setting, writer);
+                AddressSettingDefinition.REDISTRIBUTION_DELAY.marshallAsElement(setting, writer);
+                AddressSettingDefinition.SEND_TO_DLA_ON_NO_ROUTE.marshallAsElement(setting, writer);
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeConnectorServices(XMLExtendedStreamWriter writer, ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            writer.writeStartElement(Element.CONNECTOR_SERVICES.getLocalName());
+            for(final Property property : node.asPropertyList()) {
+                writer.writeStartElement(Element.CONNECTOR_SERVICE.getLocalName());
+                writer.writeAttribute(Attribute.NAME.getLocalName(), property.getName());
+                final ModelNode service = property.getValue();
+                for (AttributeDefinition attribute : ConnectorServiceDefinition.ATTRIBUTES) {
+                    attribute.marshallAsElement(property.getValue(), writer);
+                }
+                if (service.hasDefined(CommonAttributes.PARAM)) {
+                    for (Property param : service.get(CommonAttributes.PARAM).asPropertyList()) {
+                        writer.writeEmptyElement(Element.PARAM.getLocalName());
+                        writer.writeAttribute(Attribute.KEY.getLocalName(), param.getName());
+                        writer.writeAttribute(Attribute.VALUE.getLocalName(), param.getValue().get(CommonAttributes.VALUE.getName()).asString());
+                    }
+                }
+                writer.writeEndElement();
+            }
+            writer.writeEndElement();
+            writeNewLine(writer);
+        }
+    }
+
+    private static void writeConnectionFactories(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            for (Property prop : properties) {
+                final String name = prop.getName();
+                final ModelNode factory = prop.getValue();
+                if (factory.isDefined()) {
+                    writer.writeStartElement(Element.CONNECTION_FACTORY.getLocalName());
+                    writer.writeAttribute(Attribute.NAME.getLocalName(), name);
+
+                    writeRegularConnectionFactoryAttributes(writer, name, factory);
+                    writeCommonConnectionFactoryAttributes(writer, name, factory);
+                }
+            }
+        }
+    }
+
+    private static void writePooledConnectionFactories(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            for (Property prop : properties) {
+                final String name = prop.getName();
+                final ModelNode factory = prop.getValue();
+                if (factory.isDefined()) {
+                    writer.writeStartElement(Element.POOLED_CONNECTION_FACTORY.getLocalName());
+
+                    writer.writeAttribute(Attribute.NAME.getLocalName(), name);
+
+                    writePooledConnectionFactoryAttributes(writer, name, factory);
+                    writeCommonConnectionFactoryAttributes(writer, name, factory);
+                }
+            }
+        }
+    }
+
+    private static void writeRegularConnectionFactoryAttributes(XMLExtendedStreamWriter writer, String name, ModelNode factory) throws XMLStreamException {
+        ConnectionFactoryAttributes.Regular.FACTORY_TYPE.marshallAsElement(factory, writer);
+    }
+
+    private static void writeCommonConnectionFactoryAttributes(XMLExtendedStreamWriter writer, String name, ModelNode factory) throws XMLStreamException {
+        if (CommonAttributes.DISCOVERY_GROUP_NAME.isMarshallable(factory)) {
+            writer.writeStartElement(Element.DISCOVERY_GROUP_REF.getLocalName());
+            CommonAttributes.DISCOVERY_GROUP_NAME.marshallAsAttribute(factory, writer);
+            writer.writeEndElement();
+        }
+
+        // write the element for compatibility sake but it is deprecated
+        Common.DISCOVERY_INITIAL_WAIT_TIMEOUT.marshallAsElement(factory, writer);
+
+        ConnectorAttribute.CONNECTOR.marshallAsElement(factory, writer);
+
+        JndiEntriesAttribute.CONNECTION_FACTORY.marshallAsElement(factory, writer);
+
+        CommonAttributes.HA.marshallAsElement(factory, writer);
+        Common.CLIENT_FAILURE_CHECK_PERIOD.marshallAsElement(factory, writer);
+        Common.CONNECTION_TTL.marshallAsElement(factory, writer);
+        CommonAttributes.CALL_TIMEOUT.marshallAsElement(factory, writer);
+        Common.COMPRESS_LARGE_MESSAGES.marshallAsElement(factory, writer);
+        Common.CONSUMER_WINDOW_SIZE.marshallAsElement(factory, writer);
+        CommonAttributes.CONSUMER_MAX_RATE.marshallAsElement(factory, writer);
+        Common.CONFIRMATION_WINDOW_SIZE.marshallAsElement(factory, writer);
+        Common.PRODUCER_WINDOW_SIZE.marshallAsElement(factory, writer);
+        Common.PRODUCER_MAX_RATE.marshallAsElement(factory, writer);
+        Common.CACHE_LARGE_MESSAGE_CLIENT.marshallAsElement(factory, writer);
+        Common.MIN_LARGE_MESSAGE_SIZE.marshallAsElement(factory, writer);
+        CommonAttributes.CLIENT_ID.marshallAsElement(factory, writer);
+        Common.DUPS_OK_BATCH_SIZE.marshallAsElement(factory, writer);
+        Common.TRANSACTION_BATCH_SIZE.marshallAsElement(factory, writer);
+        Common.BLOCK_ON_ACKNOWLEDGE.marshallAsElement(factory, writer);
+        Common.BLOCK_ON_NON_DURABLE_SEND.marshallAsElement(factory, writer);
+        Common.BLOCK_ON_DURABLE_SEND.marshallAsElement(factory, writer);
+        Common.AUTO_GROUP.marshallAsElement(factory, writer);
+        Common.PRE_ACKNOWLEDGE.marshallAsElement(factory, writer);
+        Common.RETRY_INTERVAL.marshallAsElement(factory, writer);
+        Common.RETRY_INTERVAL_MULTIPLIER.marshallAsElement(factory, writer);
+        Common.MAX_RETRY_INTERVAL.marshallAsElement(factory, writer);
+        Common.RECONNECT_ATTEMPTS.marshallAsElement(factory, writer);
+        Common.FAILOVER_ON_INITIAL_CONNECTION.marshallAsElement(factory, writer);
+        Common.FAILOVER_ON_SERVER_SHUTDOWN.marshallAsElement(factory, writer);
+        Common.CONNECTION_LOAD_BALANCING_CLASS_NAME.marshallAsElement(factory, writer);
+        Common.USE_GLOBAL_POOLS.marshallAsElement(factory, writer);
+        Common.SCHEDULED_THREAD_POOL_MAX_SIZE.marshallAsElement(factory, writer);
+        Common.THREAD_POOL_MAX_SIZE.marshallAsElement(factory, writer);
+        Common.GROUP_ID.marshallAsElement(factory, writer);
+
+        writer.writeEndElement();
+    }
+
+
+    private static void writePooledConnectionFactoryAttributes(XMLExtendedStreamWriter writer, String name, ModelNode factory)
+            throws XMLStreamException {
+
+        Pooled.USER.marshallAsElement(factory, writer);
+        Pooled.PASSWORD.marshallAsElement(factory, writer);
+
+        if(factory.hasDefined(INBOUND_CONFIG)) {
+            final ModelNode inboundConfigs = factory.get(INBOUND_CONFIG);
+            if (inboundConfigs.getType() == ModelType.LIST) {
+                writer.writeStartElement(Element.INBOUND_CONFIG.getLocalName());
+                for (ModelNode config : inboundConfigs.asList()) {
+                    if (config.isDefined()) {
+                        Pooled.USE_JNDI.marshallAsElement(config, writer);
+                        Pooled.JNDI_PARAMS.marshallAsElement(config, writer);
+                        Pooled.USE_LOCAL_TX.marshallAsElement(config, writer);
+                        Pooled.SETUP_ATTEMPTS.marshallAsElement(config, writer);
+                        Pooled.SETUP_INTERVAL.marshallAsElement(config, writer);
+                    }
+                }
+                writer.writeEndElement();
+            }
+        }
+
+        if(factory.hasDefined(TRANSACTION)) {
+            writer.writeStartElement(Element.TRANSACTION.getLocalName());
+            writeTransactionTypeAttribute(writer, Element.MODE, factory.get(TRANSACTION));
+            writer.writeEndElement();
+        }
+
+        Pooled.MIN_POOL_SIZE.marshallAsElement(factory, writer);
+        Pooled.MAX_POOL_SIZE.marshallAsElement(factory, writer);
+        Pooled.USE_AUTO_RECOVERY.marshallAsElement(factory, writer);
+    }
+
+    private static void writeJmsQueues(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            for (Property prop : properties) {
+                final String name = prop.getName();
+                final ModelNode queue = prop.getValue();
+                if (queue.isDefined()) {
+                    writer.writeStartElement(Element.JMS_QUEUE.getLocalName());
+                    writer.writeAttribute(Attribute.NAME.getLocalName(), name);
+                    JndiEntriesAttribute.DESTINATION.marshallAsElement(queue, writer);
+                    DURABLE.marshallAsElement(queue, writer);
+                    SELECTOR.marshallAsElement(queue, writer);
+                    writer.writeEndElement();
+                }
+            }
+        }
+    }
+
+    private static void writeTopics(final XMLExtendedStreamWriter writer, final ModelNode node) throws XMLStreamException {
+        List<Property> properties = node.asPropertyList();
+        if (!properties.isEmpty()) {
+            for (Property prop : properties) {
+                final String name = prop.getName();
+                final ModelNode topic = prop.getValue();
+                if (topic.isDefined()) {
+                    writer.writeStartElement(Element.JMS_TOPIC.getLocalName());
+                    writer.writeAttribute(Attribute.NAME.getLocalName(), name);
+                    JndiEntriesAttribute.DESTINATION.marshallAsElement(topic, writer);
+                    writer.writeEndElement();
+                }
+            }
+        }
+    }
+
+
+    private void writeJmsBridge(XMLExtendedStreamWriter writer, String bridgeName, ModelNode value) throws XMLStreamException {
+        writer.writeStartElement(Element.JMS_BRIDGE.getLocalName());
+
+        if (!DEFAULT.equals(bridgeName)) {
+            writer.writeAttribute(Attribute.NAME.getLocalName(), bridgeName);
+        }
+
+        JMSBridgeDefinition.MODULE.marshallAsAttribute(value, writer);
+
+        writer.writeStartElement(SOURCE.getLocalName());
+        for (AttributeDefinition attr : JMSBridgeDefinition.JMS_SOURCE_ATTRIBUTES) {
+            attr.marshallAsElement(value, writer);
+        }
+        writer.writeEndElement();
+
+        writer.writeStartElement(TARGET.getLocalName());
+        for (AttributeDefinition attr : JMSBridgeDefinition.JMS_TARGET_ATTRIBUTES) {
+            attr.marshallAsElement(value, writer);
+        }
+        writer.writeEndElement();
+
+        for (AttributeDefinition attr : JMSBridgeDefinition.JMS_BRIDGE_ATTRIBUTES) {
+            if (attr == JMSBridgeDefinition.MODULE) {
+                // handled as a XML attribute
+                continue;
+            }
+            attr.marshallAsElement(value, writer);
+        }
+
+        writer.writeEndElement();
+    }
+
+    private static void writeTransactionTypeAttribute(final XMLExtendedStreamWriter writer, final Element attr, final ModelNode value) throws XMLStreamException {
+        writer.writeAttribute(attr.getLocalName(), value.asString());
+    }
+
+    private static void writeAttribute(final XMLExtendedStreamWriter writer, final Attribute attr, final ModelNode value) throws XMLStreamException {
+        writer.writeAttribute(attr.getLocalName(), value.asString());
+    }
+
+    private static void writeNewLine(XMLExtendedStreamWriter writer) throws XMLStreamException {
+        writer.writeCharacters(NEW_LINE, 0, 1);
+    }
+
+}
